fix: model structure adjust for 1300 issue
